Ported commit 31cf06f21889a57efc8bd5b7db637b990e4f8ce0 from castor/master
Fixed 32-bit limit bug in DiskWriteTask::checkErrors.
Refs: commit e13f841a8c21d35f1bf75dc63d8c98b21d8dd8cb in castor/master

+++ tapeserver/castor/tape/tapeserver/daemon/DiskWriteTask.cpp
@@ -206,7 +206,7 @@ void DiskWriteTask::releaseAllBlock(){
 //------------------------------------------------------------------------------  
   void DiskWriteTask::checkErrors(MemBlock* mb,int blockId,cta::log::LogContext&  lc){
     using namespace cta::log;
-    if(m_retrieveJob->retrieveRequest.archiveFileID != static_cast<unsigned int>(mb->m_fileid)
+    if(m_retrieveJob->retrieveRequest.archiveFileID != mb->m_fileid
             || blockId != mb->m_fileBlock  || mb->isFailed() ){
       LogContext::ScopedParam sp[]={
         LogContext::ScopedParam(lc, Param("received_archiveFileID", mb->m_fileid)),
